+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Added support for opting out of emitting the `@see` directives in schema doc comments through the `omitIdComments` option. By default, these tags _will_ be emitted to preserve backwards-compatibility. To omit these comments, `omitIdComments: true` can be specified as an option.
+- Introduce the `shouldOmitTypeEmit` option to the `compile` method of `Parser` instances. This option allows consumers to conditionally skip the emit of some sub-schemas' types. This function receives references to the child node being emitted and the parent node.
 
 ## [1.4.1] - 2021-04-05
 ### Fixed